Modified install script to support Hexagon SDK 3.0
This version of the SDK switches to a GCC 4.9 cross compiler.
The new cross compiler is native x86_64 and does not required the i32 libs.

The Hexagon cross compiler is 7.2.12 and is installed as part of the SDK.

The only way to change the install location of HEXAGON_Tools is to set the
HOME environment variable during the install.

The installation layout is:

${HOME}/Qualcomm/
	Hexagon_SDK/3.0
	HEXAGON_Tools/7.2.12/Tools
	ubuntu_14.04_armv7_sysroot

Signed-off-by: Mark Charlebois <charlebm@gmail.com>